+++ apis_core/apis_metainfo/templates/apis_metainfo/tags/class_definition.html
