+++ apis_core/apis_metainfo/templates/base.html
@@ -1,6 +1,6 @@
 <!DOCTYPE html>
 {% load static %}
-{% load browsing_extras %}
+{% load apis_metainfo_extras %}
 <html lang="en">
   <head>
     <meta charset="utf-8">
