+++ apis_core/apis_metainfo/templates/browsing/generic_list.html
@@ -1,7 +1,7 @@
 {% extends basetemplate %}
 {% load static %}
 {% load django_tables2 %}
-{% load browsing_extras %}
+{% load apis_metainfo_extras %}
 {% load charts_extras %}
 {% load i18n %}
 {% load crispy_forms_field %}
